requirements.txt
click==8.0.3
docker==5.0.3
grpcio==1.59.2
protobuf==4.25.1
psutil==5.9.6
